offer theme (#1236)
* move modules into assets

rm import
fix improper li nesting
break counts apparently

* create a reviews page

feel free to move around/rename

* Style tweaks

* Rename reviews to prs-needing-review

I can imagine wanting a "reviews" dashboard that displays broader
information in the future, e.g. things that have been reviewed.

* things we should really do

1. move selection to [data-attributes]
1. mark up with BEM

---------
