+++ common-theme/assets/styles/states/bad.scss
@@ -1,0 +1,5 @@
+//use for failure/incorrect/negative
+
+.is-bad {
+  background: var(--theme-color--bad);
+}
