+++ common-theme/assets/styles/states/problem.scss
@@ -1,0 +1,5 @@
+//use for warning alerts
+
+.is-problem {
+  font-weight: bolder;
+}
