--- org-cyf-itp/static/prs/needs-review/index.html
@@ -1,85 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>ITP PRs needing review</title>
-        <link href="https://fonts.googleapis.com/css2?family=Noto+Color+Emoji&amp;family=Raleway:wght@300;600;800;900&amp;family=Lato:wght@500;800&amp;display=swap" rel="stylesheet" media="all" onload="this.media=&quot;all&quot;">
-        <meta charset="UTF-8">
-        <style type="text/css">
-        body {
-            font-family: Raleway, sans-serif;
-        }
-        #overview {
-            display: grid;
-            grid-template-columns: repeat(4, 1fr);
-            gap: 40px;
-        }
-        .overview-card {
-            padding: 0px 5px 20px 5px;
-            text-align: center;
-        }
-        .age-container {
-            display: grid;
-            grid-template-columns: repeat(3, 1fr);
-        }
-        .age-bucket .count.problem {
-            font-weight: bolder;
-        }
-        .status-bad {
-            background-color: red;
-        }
-        .status-medium {
-            background-color: orange;
-        }
-        .status-good {
-            background-color: green;
-        }
-        dl {
-            display: grid;
-            grid-template-columns: max-content auto;
-        }
-
-        dt {
-            grid-column-start: 1;
-        }
-
-        dd {
-            grid-column-start: 2;
-        }
-        </style>
-    </head>
-    <body>
-        <h1>ITP PRs needing review</h1>
-        Modules in red have lots of PRs, or old PRs needing review.
-        <div id="overview"></div>
-        For PRs:
-        <dl>
-            <dt>🔴</dt><dd>Waiting for review >4 weeks</dd>
-            <dt>🟠</dt><dd>Waiting for review >1 week</dd>
-            <dt>🟢</dt><dd>From this week</dd>
-        </dl>
-        <div id="pr-list">Loading...</div>
-        <script type="module" src="index.mjs"></script>
-
-        <template class="overview-card">
-            <div class="overview-card">
-                <h3 class="module"></h2>
-                <div class="summary-heading">
-                    PRs needing review from:
-                    <div class="age-container">
-                        <div class="age-bucket this-week">This week<br /><span class="count"></span></div>
-                        <div class="age-bucket this-month">2-4 weeks ago<br /><span class="count"></span></div>
-                        <div class="age-bucket old">4+ weeks ago<br /><span class="count"></span></div>
-                    </div>
-                <div class="summary-values"></div>
-            </div>
-        </template>
-        <template class="pr-list">
-            <h2 class="module"></h2>
-            <ul class="pr-list">
-            </ul>
-        </template>
-        <template class="pr-in-list">
-            <li><span class="emoji"></span> <a class="pr-link"></a> (<a class="user-link"></a> - #<span class="pr-number"></span>)</li>
-        </template>
-    </body>
-</html>
